30/12/2022

+++ flutter_application_1/lib/Services/config_service.dart
@@ -4,7 +4,7 @@ import 'package:flutter_application_1/Models/api_response.dart';
 import 'package:http/http.dart' as http;
 
 class ConfigService {
-  Future<int> postConfig(String printing, String material) async {
+  Future<double> postConfig(String printing, String material) async {
     var response = await http.post(
       Uri.parse(APIResponse.baseUrl + APIResponse.config),
       headers: <String, String>{
@@ -17,7 +17,7 @@ class ConfigService {
     );
     if (response.statusCode == 200) {
       var result = jsonDecode(response.body);
-      int price = (result['new_price']);
+      double price = (result['new_price']);
       print(price);
       return price;
     } else {
